CCS-36 refactor: 패키지 변경, 시작 글자 완성 -> 포함 글자 완성
